+++ src/general/helpers/date_and_time/date_and_time.js
@@ -46,6 +46,7 @@ const OdsDateTime = {
         hourCycle: 'h24',
         calendar: 'gregory',
         timeZone: 'Europe/Oslo',
+
       },
     };
     dateFromOptions.type = 'from';
@@ -61,14 +62,14 @@ const OdsDateTime = {
           options.localeOptions = { hour: 'numeric', minute: 'numeric', ...options.localeOptions };
 
           if (options.type === 'to' && typeof options.prefix === 'undefined') {
-            options.prefix = '&mdash;';
+            options.prefix = '–';
           }
           break;
         case 'datetime':
-          options.localeOptions = { year: 'numeric', month: 'numeric', day: 'numeric', hour: 'numeric', minute: 'numeric', ...options.localeOptions };
+          options.localeOptions = { year: 'numeric', month: '2-digit', day: '2-digit', hour: 'numeric', minute: 'numeric', ...options.localeOptions };
 
           if (options.type === 'to' && options.format === 'time' && typeof options.prefix === 'undefined') {
-            options.prefix = '&mdash;';
+            options.prefix = '–';
           }
           break;
         case 'daytime':
@@ -78,10 +79,10 @@ const OdsDateTime = {
           options.localeOptions = { ...options.localeOptions };
           break;
         default:
-          options.localeOptions = { year: 'numeric', month: 'numeric', day: 'numeric', ...options.localeOptions };
+          options.localeOptions = { year: 'numeric', month: '2-digit', day: '2-digit', ...options.localeOptions };
 
           if (options.type === 'to' && typeof options.prefix === 'undefined') {
-            options.prefix = '&mdash;';
+            options.prefix = '–';
           }
       }
     });