+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed all dates to actually adhere to "Keep a changelog".
-- Fixed locale formatting for dates to adhere to UKEs writing rules.
+- Fixed locale formatting for dates to adhere to Oslo kommunes writing rules.
 
 ### Added