+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed all dates to actually adhere to "Keep a changelog".
+- Fixed locale formatting for dates to adhere to Oslo kommunes writing rules.
+
+### Added
+
+- Added missing developer tab for date and time.
 - Fixed an issue with the close button focus outline in modal.
 
 ## [2.0.0] - 2024-05-23