+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed all dates to actually adhere to "Keep a changelog".
+- Fixed an issue causing some links to not break correctly.
 - Fixed locale formatting for dates to adhere to Oslo kommunes writing rules.
 
 ### Added