+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added the Oslove logo.
+
 ### Fixed
 
 - Fixed all dates to actually adhere to "Keep a changelog".
 - Fixed spacing on calendar icon in date component.
+- Fixed an issue causing some links to not break correctly.
 - Fixed locale formatting for dates to adhere to Oslo kommunes writing rules.
 
 ### Added