+++ CHANGELOG.md
@@ -9,11 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added the Oslove logo.
 - Added new link-arrow icon.
 
 ### Fixed
 
 - Fixed all dates to actually adhere to "Keep a changelog".
+- Fixed an issue causing some links to not break correctly.
+- Fixed locale formatting for dates to adhere to Oslo kommunes writing rules.
+
+### Added
+
+- Added missing developer tab for date and time.
 
 ## [2.0.0] - 2024-05-23