+++ CHANGELOG.md
@@ -10,12 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the Oslove logo.
+- Added new link-arrow icon.
 
 ### Fixed
 
 - Fixed all dates to actually adhere to "Keep a changelog".
+- Fixed spacing on calendar icon in date component.
 - Fixed an issue causing some links to not break correctly.
 - Fixed locale formatting for dates to adhere to Oslo kommunes writing rules.
+- Fixed an issue with the close button focus outline in modal.
+- Fixed outline on dropdown elements in search and search seasons.
 - Fixed an issue with the close button focus outline in modal.
 
 ### Added